Fix home/away score swap on re-sync and fpts decimal truncation
Correctness fixes in sync_service from the application review:

- Re-syncing a week no longer trusts Sleeper's response order: if an
  existing matchup's stored home roster is the second team in the
  response, the pair is realigned before updating, so home/away points
  and winner can never swap between syncs.
- points_for/points_against now include Sleeper's fpts_decimal /
  fpts_against_decimal fractions instead of truncating to whole points
  (Sleeper stores 800.55 as fpts=800, fpts_decimal=55). Columns changed
  Integer -> Float with migration f7a8b9c0d1e2; season award rows keep
  integer points via explicit casts.
- roster.division is now updated on re-sync, not only at creation.
- playoff_week_start default unified to 15 (was 14 in _sync_season vs
  15 in sync_all_history).
- Sync tests no longer hit the real keeptradecut.com: an autouse
  fixture mocks refresh_ktc_values (suite runtime for test_sync
  dropped ~19s -> ~4s). Added tests for decimal preservation and
  home/away stability under reversed response order.

+++ backend/app/models/roster.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, Integer, ForeignKey, JSON
+from sqlalchemy import Column, String, Integer, Float, ForeignKey, JSON
 from sqlalchemy.orm import relationship
 from app.database import Base
 
@@ -21,8 +21,8 @@ class Roster(Base):
     wins = Column(Integer, default=0)
     losses = Column(Integer, default=0)
     ties = Column(Integer, default=0)
-    points_for = Column(Integer, default=0)
-    points_against = Column(Integer, default=0)
+    points_for = Column(Float, default=0.0)
+    points_against = Column(Float, default=0.0)
 
     # Player IDs
     players = Column(JSON)  # List of player IDs on roster
